+++ NEWS.md
@@ -1,5 +1,7 @@
 # RprobitB 1.0.0.9000
 
+* Added print method for `RprobitB_parameter`.
+
 * The function `choice_probs` is now called `choice_probabilities()` to make its functionality clearer.
 
 * Splitting the data set into a train and test part can now be done by the function `train_test()`. Consequently, the argument `test_prob` was removed in `prepare_data()` and `simulate_choices()`.